core/backend-pos/requirements.txt
# Core Framework
fastapi==0.109.2
uvicorn[standard]==0.27.1
pydantic==2.6.1
pydantic-settings==2.1.0
python-dotenv==1.0.0

# Vector Database & Embeddings
chromadb==0.4.22
sentence-transformers==2.3.1
numpy==1.26.3

# Audio Processing
openai-whisper==20231117
pydub==0.25.1
python-magic==0.4.27
python-multipart==0.0.7
noisereduce==3.0.0

# Async & Utilities
httpx==0.26.0
aiofiles==23.2.1
tenacity==8.2.3

# Logging & Monitoring
structlog==24.1.0

# Development (remove in production)
pytest==8.0.0
pytest-asyncio==0.23.4
